requirements.txt
gspread==4.0.4
oauth2client==4.1.3
python-telegram-bot==20.3
sqlalchemy==1.4.0
reportlab==3.5.67

